Split BaseService and BaseJsonService into separate modules
There’s a lot of behavior here, and going to be even more, so I think it makes sense to split these up as I’ve done with the tests.
