+++ services/appveyor/appveyor.service.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const Joi = require('joi')
-const { BaseJsonService } = require('../base')
+const BaseJsonService = require('../base-json')
 
 const appVeyorSchema = Joi.object({
   build: Joi.object({
