+++ services/base.js
@@ -1,6 +1,5 @@
 'use strict'
 
-const Joi = require('joi')
 // See available emoji at http://emoji.muan.co/
 const emojic = require('emojic')
 const chalk = require('chalk')
@@ -12,7 +11,6 @@ const {
   makeColor,
   setBadgeColor,
 } = require('../lib/badge-data')
-const { checkErrorResponse, asJson } = require('../lib/error-helper')
 // Config is loaded globally but it would be better to inject it. To do that,
 // there needs to be one instance of the service created at registration time,
 // which gets the config injected into it, instead of one instance per request.
@@ -308,59 +306,4 @@ class BaseService {
   }
 }
 
-class BaseJsonService extends BaseService {
-  static _validate(json, schema) {
-    const { error, value } = Joi.validate(json, schema, {
-      allowUnknown: true,
-      stripUnknown: true,
-    })
-    if (error) {
-      this.logTrace(
-        'error',
-        emojic.womanShrugging,
-        'Response did not match schema',
-        error.message
-      )
-      throw new InvalidResponse({
-        prettyMessage: 'invalid json response',
-        underlyingError: error,
-      })
-    } else {
-      this.logTrace('validate', emojic.bathtub, 'JSON after validation', value)
-      return value
-    }
-  }
-
-  async _requestJson({ schema, url, options = {}, notFoundMessage }) {
-    const logTrace = (...args) => this.constructor.logTrace('fetch', ...args)
-    if (!schema || !schema.isJoi) {
-      throw Error('A Joi schema is required')
-    }
-    const mergedOptions = {
-      ...{ headers: { Accept: 'application/json' } },
-      ...options,
-    }
-    logTrace(emojic.bowAndArrow, 'Request', url, '\n', mergedOptions)
-    return this._sendAndCacheRequest(url, mergedOptions)
-      .then(({ res, buffer }) => {
-        logTrace(emojic.dart, 'Response status code', res.statusCode)
-        return { res, buffer }
-      })
-      .then(
-        checkErrorResponse.asPromise(
-          notFoundMessage ? { notFoundMessage: notFoundMessage } : undefined
-        )
-      )
-      .then(asJson)
-      .then(json => {
-        logTrace(emojic.dart, 'Response JSON (before validation)', json)
-        return json
-      })
-      .then(json => this.constructor._validate(json, schema))
-  }
-}
-
-module.exports = {
-  BaseService,
-  BaseJsonService,
-}
+module.exports = BaseService
