+++ services/gem/gem-downloads.service.js
@@ -3,7 +3,7 @@
 const semver = require('semver')
 const Joi = require('joi')
 
-const { BaseJsonService } = require('../base')
+const BaseJsonService = require('../base-json')
 const { InvalidResponse } = require('../errors')
 const {
   downloadCount: downloadCountColor,
