+++ services/gem/gem-rank.service.js
@@ -2,7 +2,7 @@
 
 const Joi = require('joi')
 
-const { BaseJsonService } = require('../base')
+const BaseJsonService = require('../base-json')
 const { floorCount: floorCountColor } = require('../../lib/color-formatters')
 const { ordinalNumber } = require('../../lib/text-formatters')
 const { nonNegativeInteger } = require('../validators.js')
