+++ services/gem/gem-version.service.js
@@ -2,7 +2,7 @@
 
 const Joi = require('joi')
 
-const { BaseJsonService } = require('../base')
+const BaseJsonService = require('../base-json')
 const { addv: versionText } = require('../../lib/text-formatters')
 const { version: versionColor } = require('../../lib/color-formatters')
