+++ services/npm/npm-downloads.service.js
@@ -1,7 +1,7 @@
 'use strict'
 
 const Joi = require('joi')
-const { BaseJsonService } = require('../base')
+const BaseJsonService = require('../base-json')
 const { metric } = require('../../lib/text-formatters')
 const { nonNegativeInteger } = require('../validators.js')
